Minor Update
Changed location of time classes to extensions and renamed to 'Clock'
